+++ src/main/java/com/example/gak/domain/member/service/MemberCommandService.java
@@ -18,8 +18,8 @@ import com.example.gak.domain.record.repository.RecordRepository;
 import com.example.gak.domain.session.entity.enums.SessionRoomStatus;
 import com.example.gak.domain.session.repository.SessionRoomMemberRepository;
 import com.example.gak.domain.session.repository.SessionRoomRepository;
-import com.example.gak.domain.subtask.repository.SubTaskRepository;
 import com.example.gak.domain.task.entity.Task;
+import com.example.gak.domain.task.repository.SubTaskRepository;
 import com.example.gak.domain.task.repository.TaskRepository;
 import com.example.gak.global.apiPayload.code.GeneralErrorCode;
 import com.example.gak.global.apiPayload.exception.GeneralException;
