+++ src/main/java/com/example/gak/domain/session/service/SessionCommandService.java
@@ -19,9 +19,9 @@ import com.example.gak.domain.session.entity.enums.SessionParticipantRole;
 import com.example.gak.domain.session.entity.enums.SessionRoomStatus;
 import com.example.gak.domain.session.repository.SessionRoomMemberRepository;
 import com.example.gak.domain.session.repository.SessionRoomRepository;
-import com.example.gak.domain.subtask.entity.SubTask;
-import com.example.gak.domain.subtask.repository.SubTaskRepository;
+import com.example.gak.domain.task.entity.SubTask;
 import com.example.gak.domain.task.entity.Task;
+import com.example.gak.domain.task.repository.SubTaskRepository;
 import com.example.gak.domain.task.repository.TaskRepository;
 import com.example.gak.global.apiPayload.code.GeneralErrorCode;
 import com.example.gak.global.apiPayload.exception.GeneralException;
