+++ src/main/java/com/example/gak/domain/task/repository/SubTaskRepository.java
@@ -1,4 +1,4 @@
-package com.example.gak.domain.subtask.repository;
+package com.example.gak.domain.task.repository;
 
 import java.util.List;
 
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.example.gak.domain.subtask.entity.SubTask;
+import com.example.gak.domain.task.entity.SubTask;
 
 public interface SubTaskRepository extends JpaRepository<SubTask, Long> {
